south-indian-jewelry/backend/requirements.txt
flask==3.0.0
flask-pymongo==2.3.0
pymongo==4.6.1
flask-cors==4.0.0
flask-bcrypt==1.0.1
flask-session==0.5.0
python-dotenv==1.0.0
